resolve a circular reference, fixes #20
When the history subscriber of the bundle is enabled and a firewall
uses an entity user provider, there is a circular reference. Now, the
subscriber no longer makes use of Doctrine's event system. Instead,
the bundle creates its own event subsystem. Events are now dispatched
by the translation manager implementations of the database driver to
which the history subscriber is listening.